\docType{data}
\name{abortion}
\alias{abortion}
\title{Haberman's abortion, education, and religion data}
\format{A 3x3x3 (contingency) table}
\usage{
data(abortion)
}
\description{
Haberman's abortion, education, and religion data
}
\author{
1972 National Opinion Research Center
}
\references{
Haberman, S. (1978). \emph{Analysis of Qualitative Data} 1,
2. Academic Press, Orlando, FL.
}
\keyword{datasets}

