\docType{package}
\name{algeo}
\alias{algeo}
\alias{algeo-package}
\alias{package-algeo}
\title{algeo : Symbolic and Numerical Algebraic Geometry in R}
\description{
The algeo package is intended to allow R users to access
the power of the algebraic geometry software Macaulay2 and
Bertini.  Using structures from the mpoly package, algeo
provides convenience functions for finding solution sets as
well as low-level functions to pass code to Macaulay2 and
Bertini.
}

