% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makers.R
\name{Amaker}
\alias{Amaker}
\title{Distance transitive matrix}
\usage{
Amaker(m, k)
}
\arguments{
\item{m}{the number of objects}

\item{k}{the number of objects selected}
}
\value{
...
}
\description{
Compute the distance transitive matrix for a survey in which k objects are
selected from a group of m
}
\examples{


Amaker(4, 2)

}
\seealso{
\code{\link[=Tmaker]{Tmaker()}}, \code{\link[=Emaker]{Emaker()}}, \code{\link[=Mmaker]{Mmaker()}}, \code{\link[=Pmaker]{Pmaker()}}, \code{\link[=Smaker]{Smaker()}}
}
