% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apoisson.R
\name{apoisson}
\alias{apoisson}
\title{Fit a hierarchical Poisson regression models with algebraic
methods}
\usage{
apoisson(model, data, init = tab2vec(data), iter = 10000, burn = 1000,
  thin = 10, engine = c("Cpp", "R"), method = c("ipf", "mcmc"), moves,
  ...)
}
\arguments{
\item{model}{hierarchical log-linear model specification}

\item{data}{data, typically as a table but can be in different 
formats.  see \code{\link{teshape}}}

\item{init}{the initialization of the chain. by default, this is 
the observed table}

\item{iter}{number of chain iterations}

\item{burn}{burn-in}

\item{thin}{thinning}

\item{engine}{C++ or R? (C++ yields roughly a 20-25x speedup)}

\item{method}{should the expected value (exp) be fit using 
iterative proportional fitting (via loglin) or the MCMC as the 
average of the steps?}

\item{moves}{the markov moves for the mcmc (as columns of a 
matrix).}

\item{...}{...}
}
\value{
a list containing named elements \itemize{ \item 
  \code{steps}: an integer matrix whose columns represent 
  individual samples from the mcmc. \item \code{moves}: the moves
  used for the proposal distribution in the mcmc, computed with 
  4ti2 (note that only the positive moves are given). \item 
  \code{acceptProb}: the average acceptance probability of the 
  moves, including the thinned moves. \item \code{param}: the 
  fitted parameters of the log linear model. \item \code{df}: 
  parameters per term in the model \item \code{quality}: model 
  selection statistics AIC, AICc, and BIC. \item 
  \code{residuals}: the (unstandardized) pearson residuals (O - 
  E) / sqrt(E) \item \code{call}: the call. \item \code{obs}: the
  contingency table given. \item \code{exp}: the fit contingency 
  table as an integer array. \item \code{A}: the sufficient 
  statistics computing matrix (from Tmaker). \item 
  \code{p.value}: the exact p-values of individual tests, 
  accurate to Monte-Carlo error.  these are computed as the 
  proportion of samples with statistics equal to or larger than 
  the oberved statistic. \item \code{mid.p.value}: the mid 
  p.values, see Agresti pp.20--21. \item \code{statistic}: the 
  pearson's chi-squared (X2), likelihood ratio (G2), 
  Freeman-Tukey (FT), Cressie-Read (CR), and Neyman modified 
  chi-squared (NM) statistics computed for the table given. \item
  \code{sampsStats}: the statistics computed for each mcmc 
  sample. \item \code{cells}: the number of cells in the table. 
  \item \code{method}: the method used to estimate the table. }
}
\description{
Fit a hierarchical Poisson regression models with algebraic
methods
}
\examples{

\dontrun{


## handedness introductory example
############################################################

data(handy)

(out <- loglinear(~ Gender + Handedness, data = handy))

# you can also specify the same model using variable indices...   
(out <- loglinear(~ 1 + 2, data = handy))
  
# ... or as a list of facets given by indices
(out <- loglinear(list(1, 2), data = handy))

# ... or as a list of facets given by name
(out <- loglinear(list("Gender", "Handedness"), data = handy))

# ... and even via a pre-computed configuration matrix
# this method does come with somewhat reduced output
A <- hmat(c(2, 2), 1:2)
(out <- loglinear(A, data = handy))



# loglinear performs the same tasks as loglin and loglm,
# but loglinear gives the exact test p values and more goodness-of-fit statistics
stats::loglin(handy, list(1, 2))
MASS::loglm(~ Gender + Handedness, data = handy)
# loglm is just a wrapper of loglin  
  
# we can check loglinear's output with
fisher.test(handy)$p.value
out$p.value



  






}


}
\author{
David Kahle
}
\references{
Diaconis, P. and B. Sturmfels (1998). Algebraic 
  Algorithms for Sampling from Conditional Distributions. 
  \emph{The Annals of Statistics} 26(1), pp.363-397.

Drton, M., B. Sturmfels, and S. Sullivant (2009). 
  \emph{Lectures on Algebraic Statistics}, Basel: Birkhauser 
  Verlag AG.

Aoki, S., H. Hara, and A. Takemura (2012). 
  \emph{Markov Bases in Algebraic Statistics}, Springer.

Agresti, A. (2002). \emph{Categorical Data Analysis},
  Basel: John Wiley & Sons, 2ed.

Agresti, A. (1992). A Survey of Exact Inference for 
  Contingency Tables \emph{Statistical Science} 7(1), pp.131-153.

Read, T. and Cressie, N. (1998). 
  \emph{Goodness-of-Fit Statistics for Discrete Multivariate 
  Data}, Springer-Verlag.
}
\seealso{
\code{\link{loglin}}, \code{\link{loglm}}, 
  \code{\link{metropolis}}
}

