% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab2array.r
\name{array2tab}
\alias{array2tab}
\title{Array to table conversion}
\usage{
array2tab(array)
}
\arguments{
\item{array}{an array}
}
\value{
an tabl
}
\description{
Convert an array into a table.
}
\examples{

mat <- matrix(rpois(25,2), 5, 5) 
array2tab(mat)

data(handy)
array2tab(handy)


}
\seealso{
\code{\link{tab2array}}
}

