% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthdeath.R
\docType{data}
\name{birthdeath}
\alias{birthdeath}
\title{Andrews/Herzberg Birthday and Deathday Dataset}
\format{
A 12x12 (contingency) table
}
\usage{
data(birthdeath)
}
\description{
A two-way table cross-classifying birth and death days of 82 individuals.
}
\references{
Andrews, D. and A. Herzberg (1985). \emph{Data}. Springer-Verlag,
New York.
}
\keyword{datasets}
