% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_bounds.R
\name{cell-bounds}
\alias{cell-bounds}
\alias{cell_bounds}
\alias{cellBounds}
\title{Compute the lower and upper bounds}
\usage{
cell_bounds(A, b, strategy = c("lp", "ip"), start = 1, messaging = FALSE)

cellBounds(...)
}
\arguments{
\item{A}{the configuration matrix of the model defining the fiber}

\item{b}{solution vector in Ax = b (vector of sufficient statistics)}

\item{strategy}{"lp" for linear programming and "ip" for integer programming.
(both use \code{\link[lpSolve:lp]{lpSolve::lp()}})}

\item{start}{reduce the size of the matrix A, i.e. use
\code{A[,start:ncol(A)]} instead of A}

\item{messaging}{TRUE for messages}

\item{...}{...}
}
\value{
a named numeric vector
}
\description{
Compute the lower and upper bounds of a cell in a vectorized contingency
table.
}
\examples{


A <- hmat(c(2,2), 1:2)
b <- rep.int(4, 4)
cell_bounds(A, b, messaging = TRUE)



}
\author{
Ruriko Yoshida \email{ruriko.yoshida@uky.edu}, David Kahle
\email{david@kahle.io}
}
