% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellBound.R
\name{cellBound}
\alias{cellBound}
\title{Compute the lower or upper bound}
\usage{
cellBound(A, b, side = c("lower", "upper"), strategy = c("lp", "ip"))
}
\arguments{
\item{A}{the configuration matrix of the model defining the fiber}

\item{b}{solution vector in Ax = b (vector of sufficient
statistics)}

\item{side}{bound side to compute, "lower" or "upper"}

\item{strategy}{"lp" for linear programming and "ip" for integer
programming. (both use \code{\link{lp}}.)}
}
\value{
an integer
}
\description{
Compute the lower or upper bound of a cell in a vectorized
contingency table.
}
\examples{


A <- hmat(c(2,2), 1:2)
b <- rep.int(4, 4)
cellBound(A, b, "lower") # 0
cellBound(A, b, "upper") # 4




}
\author{
Ruriko Yoshida \email{ruriko.yoshida@uky.edu}, David
  Kahle \email{david.kahle@gmail.com}
}

