% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cellBounds.R
\name{cellBounds}
\alias{cellBounds}
\title{Compute the lower and upper bounds}
\usage{
cellBounds(A, b, strategy = c("lp", "ip"), start = 1, messaging = FALSE)
}
\arguments{
\item{A}{the configuration matrix of the model defining the fiber}

\item{b}{solution vector in Ax = b (vector of sufficient statistics)}

\item{strategy}{"lp" for linear programming and "ip" for integer programming.
(both use \code{\link{lp}})}

\item{start}{reduce the size of the matrix A, i.e. use A[,start:ncol(A)] instead of A}
}
\value{
a named numeric vector
}
\description{
Compute the lower and upper bounds of a cell in a vectorized contingency table.
}
\examples{
A <- hmat(c(2,2), 1:2)
b <- rep.int(4, 4)
cellBounds(A, b, messaging = TRUE)
}
\author{
Ruriko Yoshida \email{ruriko.yoshida@uky.edu}, David Kahle \email{david.kahle@gmail.com}
}

