\docType{data}
\name{cookie}
\alias{cookie}
\title{Girl Scout cookie preferences}
\format{A list with elements "freq", "raw", "sorted", and "cookies"}
\usage{
data(cookie)
}
\description{
A small approval dataset on Girl Scout cookies
}
\author{
Ann Johnston and Michael Orrison
}
\references{
Johnston, A. and Orrison, M. (2011) Markov Bases for
Noncommutative Harmonic Analysis of Partially Ranked Data.
Undergraduate Thesis, Department of Mathematics, Harvey
Mudd College.
}
\keyword{datasets}

