\docType{data}
\name{drugs}
\alias{drugs}
\title{Use of alcohol, cigarettes, and marijuana in high school students}
\format{A 2x2x2 (contingency) table}
\usage{
data(drugs)
}
\description{
Use of alcohol, cigarettes, and marijuana in high school
students
}
\author{
Wright State University School of Medicine and the United
Health Services in Dayton, Ohio
}
\references{
Agresti, A. (2002). \emph{Categorical Data Analysis},
Basel: John Wiley & Sons, 2ed. (p.322)
}
\keyword{datasets}

