% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makers.r
\name{Emaker}
\alias{Emaker}
\title{Create the expected higher-order statistics calculating matrix for approval data}
\usage{
Emaker(m, vin, vout)
}
\arguments{
\item{m}{the number of objects}

\item{vin}{the (lower order) grouping level of the data}

\item{vout}{the desired higher order grouping level}
}
\value{
...
}
\description{
Create the expected higher-order statistics calculating matrix for approval data
}
\examples{
Emaker(6, 0, 1)
Emaker(6, 0, 2)
Emaker(6, 0, 3)
Emaker(6, 0, 4)

Emaker(6, 1, 1)
Emaker(6, 1, 2)
Emaker(6, 1, 3)
Emaker(6, 1, 4)
Emaker(6, 1, 5)
Emaker(6, 1, 6)

# compare to Tmaker
Emaker(6, 1, 3) # contributors when bumping up from 1-groups to 3-groups
Tmaker(6, 3, 1)
}
\seealso{
\code{\link{Tmaker}}, \code{\link{Amaker}}, \code{\link{Mmaker}}, \code{\link{Pmaker}}, \code{\link{Smaker}}
}

