% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvariety.R
\name{ggvariety}
\alias{ggvariety}
\title{Plot a variety}
\usage{
ggvariety(mp, xlim = c(-1, 1), ylim = c(-1, 1), n = 101, nx = n,
  ny = n, ...)
}
\arguments{
\item{mp}{an mpoly object}

\item{xlim}{vector representing x bounds}

\item{ylim}{vector representing y bounds}

\item{n}{number of mesh points in each dimension}

\item{nx}{number of mesh points in the abcissa (x)}

\item{ny}{number of mesh points in the ordinate (y)}

\item{...}{additional parameters}
}
\value{
A ggplot object containing variety plot
}
\description{
Uses geom_contour() and ggplot() to plot an mpoly object representing a
variety in 2D space.
}
\examples{

## basic usage
##################################################

ggvariety(mp("y - x"))
ggvariety(mp("y - x^2"))
ggvariety(mp("x^2 + y^2 - 1"))


## setting limits
##################################################

ggvariety(mp("y - x^2"))
ggvariety(mp("y - x^2"), xlim = c(-2,2), ylim = c(-2,2))


## ggplot2 styling
##################################################

library("ggplot2")
ggvariety(mp("x^2 + y^2 - 1")) + coord_equal()
ggvariety(mp("x^2 + y^2 - 1")) + coord_equal() + theme_bw()
ggvariety(mp("x^2 + y^2 - 1")) + coord_equal() + theme_void()



## possible issues
##################################################

# at a low level, ggvariety() uses grDevices::contourLines()
# to numerically detect zero crossings. this is an imperfect process,
# so you may see gaps where none exist. as a general strategy, upping
# the number of sampled points on the grid is recommended.
ggvariety("y^2 - x^3 - x^2") + coord_equal()
ggvariety("y^2 - x^3 - x^2", n = 201) + coord_equal()

}
\author{
Phillip Hossu, Ryan Hebdon, Chong Sun, Grant Innerst, David Kahle
}
