% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graver.R
\name{graver}
\alias{graver}
\alias{memGraver}
\title{Compute a Graver basis with 4ti2}
\usage{
graver(mat, format = c("mat", "vec", "tab"), dim = NULL, all = FALSE,
  dir = tempdir(), opts = "-p=gmp", quiet = TRUE, dbName)

memGraver(...)
}
\arguments{
\item{mat}{a matrix; for example the output of \code{\link{hmat}}}

\item{format}{how the moves should be returned (if "mat", moves are columns)}

\item{dim}{the dimension to be used in vec2tab if format = "tab" is used,
oftentimes a vector of the number of levels of each variable in order}

\item{all}{if TRUE, all moves (+ and -) are given.  if FALSE, only the +
moves are given}

\item{dir}{directory to place the files in, without an ending /}

\item{opts}{options for graver}

\item{quiet}{show 4ti2 output}

\item{dbName}{the name of the model in the Markov bases database,
http://markov-bases.de, see examples}

\item{...}{...}
}
\value{
a matrix containing the Graver basis as its columns (for easy
  addition to tables)
}
\description{
A Graver basis of a matrix A is computed with the graver function of 4ti2,
obtained with the LattE-integrale bundle.
}
\examples{
\dontrun{
# these examples require having extra software installed




# 2x2 independence example
# following convention, the first index indicates rows
varlvls <- c(2,2)
facets <- list(1,2)
( A <- hmat(varlvls, facets) )
graver(A)
graver(A, "vec")
graver(A, "tab", varlvls)
graver(A, "tab", varlvls, all = TRUE)




# 3x3 independence example
# following convention, the first index indicates rows
varlvls <- c(3,3)
facets <- list(1,2)
( A <- hmat(varlvls, facets) )
markov(A)
graver(A)
graver(A, "vec")
graver(A, "tab", varlvls)
graver(A, "tab", varlvls, TRUE)






# markov basis vs graver basis
varlvls <- c(3,3,3)
facets <- list(c(1,2), c(1,3), c(2,3))
( A <- hmat(varlvls, facets) )
str(zbasis(A)) #   8 elements
str(markov(A)) #  81 elements
str(graver(A)) # 795 elements






# LAS example 1.2.12, p.16  (no 3-way interaction)
varlvls <- c(2,2,2,2)
facets <- list(c(1,2), c(1,4), c(2,3))
( A <- hmat(varlvls, facets) )
graver(A)
graver(A, "tab", varlvls) # hard to understand
# tableau(graver(A), varlvls) # error?





# memoising graver - faster recomputing times
# (memorizing computed graver bases)
A <- hmat(c(3,3,2), 1:3)
system.time(graver(A))
system.time(graver(A))
system.time(memGraver(A))
system.time(memGraver(A))













markov(diag(1, 10))
graver(diag(1, 10))
graver(diag(1, 10), "vec")
graver(diag(1, 10), "vec", all = TRUE)

}
}
\references{
Drton, M., B. Sturmfels, and S. Sullivant (2009). \emph{Lectures
  on Algebraic Statistics}, Basel: Birkhauser Verlag AG.
}

