% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/groebner.r
\name{groebner}
\alias{groebner}
\title{Compute a Grobner basis with 4ti2}
\usage{
groebner(mat, format = c("mat", "vec", "tab"), dim = NULL, all = FALSE,
  dir = tempdir(), opts = "-parb", quiet = TRUE)
}
\arguments{
\item{mat}{a matrix; for example the output of \code{\link{hmat}}}

\item{format}{how the moves should be returned (if "mat", moves are columns)}

\item{dim}{the dimension to be used in vec2tab if format = "tab" is used, oftentimes a vector of the number of levels of each variable in order}

\item{all}{if TRUE, all moves (+ and -) are given.  if FALSE, only the + moves are given}

\item{dir}{directory to place the files in, without an ending /}

\item{opts}{options for groebner}

\item{quiet}{show 4ti2 output}
}
\value{
a matrix containing the Grobner basis as its columns (for easy addition to tables)
}
\description{
A Grobner basis of a matrix A is computed with the groebner function of 4ti2, obtained with the LattE-integrale bundle.
}
\examples{
\dontrun{
# these examples require having extra software installed




# 2x2 independence example
# following convention, the first index indicates rows
varlvls <- c(2,2)
facets <- list(1,2)
( A <- hmat(varlvls, facets) )
groebner(A)
groebner(A, "vec")
groebner(A, "tab", varlvls)
groebner(A, "tab", varlvls, all = TRUE)
groebner(A, quiet = FALSE)


}
}
\references{
Drton, M., B. Sturmfels, and S. Sullivant (2009). \emph{Lectures on Algebraic Statistics}, Basel: Birkhauser Verlag AG.
}
\seealso{
\code{\link{markov}}
}

