\docType{data}
\name{haberman}
\alias{haberman}
\title{Haberman's positive margin no three-way interaction MLE}
\format{A 2x2x2 (contingency) table}
\usage{
data(haberman)
}
\description{
Haberman's positive margin no three-way interaction MLE
}
\author{
unknown
}
\references{
Haberman, S. (1974). \emph{The Analysis of Frequency Data}
1, 2. University of Chicago Press, Chicago, IL.
}
\keyword{datasets}

