% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is.bertini}
\alias{is.bertini}
\title{Bertini Object Check}
\usage{
is.bertini(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
Vector of logicals.
}
\description{
Test whether an object is an bertini object.
}
\examples{
# see ?bertini
}

