% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is.linear}
\alias{is.linear}
\title{Test whether an mpoly object is linear.}
\usage{
is.linear(x)
}
\arguments{
\item{x}{an mpoly or mpolyList object}
}
\value{
a logical vector
}
\description{
Test whether an mpoly object is linear.
}
\examples{
\dontrun{

is.linear(mp("0"))
is.linear(mp("x + 1"))
is.linear(mp("x + y"))
is.linear(mp(c("0", "x + y")))

is.linear(mp("x + x y"))
is.linear(mp(c("x + x y", "x")))


}
}

