% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is.m2}
\alias{is.m2}
\title{Macaulay2 Object Check}
\usage{
is.m2(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
Vector of logicals.
}
\description{
Test whether an object is an m2 object.
}
\examples{
\dontrun{

is.m2(m2("13^1000"))

}
}

