% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/latteFiles.r
\name{latteFiles}
\alias{latteFiles}
\alias{read.latte}
\alias{write.latte}
\title{Read/write a latte file to/from disk}
\usage{
write.latte(mat, file)

read.latte(file, format = c("mat", "Ab"))
}
\arguments{
\item{mat}{a matrix}

\item{file}{the name of the file}
}
\value{
an invisible form of the saved output.
}
\description{
\code{write.latte} writes a latte-formatted file to file, and
\code{read.latte} reads a latte-formatted file from disk.
}
\examples{
\dontrun{

(mat <- matrix(sample(9), 3, 3))
write.latte(mat, "foo.hrep")
file.show("foo.hrep")
read.latte("foo.hrep")
read.latte("foo.hrep", "Ab")

attr(mat, "linearity") <- c(1, 3)
attr(mat, "nonnegative") <- 2
mat
write.latte(mat, "foo.hrep")
file.show("foo.hrep")
read.latte("foo.hrep")

file.remove("foo.hrep")

}
}
\seealso{
\code{\link{write.latte}}
}

