% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/latteOptim.r
\name{latteOptim}
\alias{latteMax}
\alias{latteMin}
\alias{latteOptim}
\title{Solve an integer progam with LattE}
\usage{
latteOptim(objective, constraints, type = c("max", "min"), method = c("lp",
  "cones"), dir = tempdir(), opts = "", quiet = TRUE)

latteMax(objective, constraints, method = c("lp", "cones"), dir = tempdir(),
  opts = "", quiet = TRUE)

latteMin(objective, constraints, method = c("lp", "cones"), dir = tempdir(),
  opts = "", quiet = TRUE)
}
\arguments{
\item{objective}{a linear polynomial to pass to \code{\link{mp}}, see
examples}

\item{constraints}{a collection of linear polynomial (in)equalities that
define the feasibility region, the integers in the polytope}

\item{method}{method LP or cones}

\item{dir}{directory to place the files in, without an ending /}

\item{opts}{options; see the LattE manual at
\url{http://www.math.ucdavis.edu/~latte}}

\item{quiet}{show latte output}
}
\value{
a named list with components \code{par}, a named-vector of optimizing
  arguments, and \code{value}, the value of the objective function at the
  optimial point
}
\description{
\code{latteMax} and \code{latteMin} use LattE's \code{latte-maximize} and
\code{latte-minimize} functions to find the maximum or minimum of a linear
objective function over the integers points in a polytope (i.e. satisfying
linearity constraints). This makes use of the digging algorithm; see the
LattE manual at \url{http://www.math.ucdavis.edu/~latte} for details.
}
\examples{
\dontrun{ # user must have LattE installed



latteMax("-2 x + 3 y", c("x + y <= 10", "x >= 0", "y >= 0"))


df <- expand.grid(x = 0:10, y = 0:10)
df <- subset(df, x + y <= 10)
df$objective <- with(df, -2*x + 3*y)
df[which.max(df$objective),]
df[which.min(df$objective),]

library(ggplot2)
qplot(x, y, data = df, size = objective)

latteMax("-2 x + 3 y", c("x + y <= 10", "x >= 0", "y >= 0"))
latteMin("-2 x + 3 y", c("x + y <= 10", "x >= 0", "y >= 0"))




latteMin("-2 x + 3 y", c("x + y <= 10", "x >= 0", "y >= 0"), method = "cones")








latteMin("-2 x - 3 y - 4 z", c(
  "3 x + 2 y + z <= 10",
  "2 x + 5 y + 3 z <= 15",
  "x >= 0", "y >= 0", "z >= 0"
), "cones", quiet = FALSE)

df <- expand.grid(x = 0:10, y = 0:10, z = 0:10)
df <- subset(df,
  (3*x + 2*y + 1*z <= 10) &
  (2*x + 5*y + 3*z <= 15)
)

df$objective <- with(df, -2*x + -3*y + -4*z)
df[which.min(df$objective),]



}
}

