% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lawrence.R
\name{lawrence}
\alias{lawrence}
\title{Compute the Lawrence lifting of a configuration matrix}
\usage{
lawrence(A, k = 1, extent = c("short", "tall"))
}
\arguments{
\item{A}{the configuration matrix of the model}

\item{k}{kth Lawrence lifting}

\item{extent}{short or tall version}
}
\value{
an integer matrix
}
\description{
Compute the Lawrence lifting of a configuration matrix
}
\examples{


(A <- hmat(c(2,2), 1:2))
plot_matrix(A) # in package latte

lawrence(A)
plot_matrix(lawrence(A))

lawrence(A, 2)
plot_matrix(lawrence(A, 2))

lawrence(A, extent = "tall")
plot_matrix(lawrence(A, extent = "tall"))

plot_matrix(A)
plot_matrix(lawrence(A))
plot_matrix(lawrence(A, 2))
plot_matrix(lawrence(A, 3))
plot_matrix(lawrence(A, 20))

plot_matrix(lawrence(A))
plot_matrix(lawrence(A, extent = "tall"))


\dontrun{ # requires 4ti2

J <- 10 # number of levels
A <- rbind(1L, 1:J)
markov(A)

lawrence(A)
plotMatrix(lawrence(A))
markov(lawrence(A))
# markov(lawrence(A, 2)) # takes unknown long time when J=10
zbasis(lawrence(A))

}


}
\seealso{
Aoki, Hara, and Takemura (2012), pp. 61-62
}
\author{
David Kahle \email{david.kahle@gmail.com}
}
