% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{lpnorm}
\alias{lpnorm}
\title{Lp Norm}
\usage{
lpnorm(x, p = 2)
}
\arguments{
\item{x}{x}

\item{p}{p}
}
\value{
...
}
\description{
Compute the Lp norm of a vector.
}
\examples{

lpnorm(1:10)
lpnorm(matrix(1:25, 5, 5))
lpnorm(split(1:25, rep(1:5, each = 5)))

lpnorm(1:10, 1)
lpnorm(matrix(1:25, 5, 5), 1)
lpnorm(split(1:25, rep(1:5, each = 5)), 1)

lpnorm(rnorm(10), 0)
lpnorm(matrix(rnorm(25), 5, 5), 0)
lpnorm(split(rnorm(25), rep(1:5, each = 5)), 0)

lpnorm(-5:5, Inf)
lpnorm(matrix(-25:-1, 5, 5), Inf)
lpnorm(split(-25:-1, rep(1:5, each = 5)), Inf)

}

