% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{mchoose}
\alias{mchoose}
\title{Multinomial Coefficient}
\usage{
mchoose(n, x)
}
\arguments{
\item{n}{an integer}

\item{x}{a vector of integers}
}
\value{
...
}
\description{
Compute the multinomial coefficient.
}
\details{
This function computes the multinomial coefficient by computing the factorial of each number on a log scale, differencing log(n!) - sum(log(x!)), and then exponentiating.  It then checks to see if this is an integer; if it's not, it issues a warning.
}
\examples{

mchoose(6, c(2,2,1,1))




}

