% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metropolis.r
\name{metropolis}
\alias{metropolis}
\alias{rawMetropolis}
\title{The Metropolis Algorithm}
\usage{
metropolis(init, moves, iter = 1000L, burn = 0L, thin = 1L,
  dist = c("hypergeometric", "uniform"), engine = c("C++", "R"),
  hit_and_run = FALSE, adaptive = FALSE)

rawMetropolis(init, moves, iter = 1000, dist = "hypergeometric",
  hit_and_run = FALSE, adaptive = FALSE)
}
\arguments{
\item{init}{the initial step}

\item{moves}{the moves to be used (the negatives will be added);
they are arranged as the columns of a matrix.}

\item{iter}{number of chain iterations}

\item{burn}{burn-in}

\item{thin}{thinning}

\item{dist}{steady-state distribution; "hypergeometric" (default)
or "uniform"}

\item{engine}{\code{C++} or \code{R}? (C++ is significantly faster)}

\item{hit_and_run}{Whether or not to use the discrete hit and run algorithm in
the metropolis algorithm. Defaulted to FALSE}

\item{adaptive}{Option when hit_and_run = TRUE. If adaptive = TRUE, hit and run will choose a proposal distribution adaptively.
Defaulted to FALSE}
}
\value{
a list containing named elements

\itemize{

\item \code{steps}: an integer matrix whose columns represent individual
samples from the mcmc.

\item \code{moves}: the moves supplied.

\item \code{accept_prob}: the empirical transition probability of the
moves, including the thinned moves.

\item \code{accept_count}: the numerator of \code{accept_prob}.

\item \code{total_count}: the numerator of \code{total_prob}.

}
}
\description{
Given a starting table (as a vector) and a collection of moves, run the
Metropolis-Hastings algorithm starting with the starting table.
}
\details{
See Algorithm 1.1.13 in LAS, the reference below.
}
\references{
Drton, M., B. Sturmfels, and S. Sullivant (2009). \emph{Lectures
on Algebraic Statistics}, Basel: Birkhauser Verlag AG.
}
\author{
David Kahle
}
