% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makers.R
\name{Mmaker}
\alias{Mmaker}
\title{Marginals matrix}
\usage{
Mmaker(m)
}
\arguments{
\item{m}{the number of objects}
}
\value{
...
}
\description{
Compute the marginals matrix for a full ranking of m objects
}
\details{
This is the transpose of the marginals matrix presented in Marden (1995).
}
\examples{


data(city); city

Mmaker(3)
Mmaker(3) \%*\% city

}
\references{
Marden, J. I. (1995). \emph{Analyzing and Modeling Rank Data},
London: Chapman & Hall. p.42.
}
\seealso{
\code{\link[=Tmaker]{Tmaker()}}, \code{\link[=Amaker]{Amaker()}}, \code{\link[=Emaker]{Emaker()}}, \code{\link[=Pmaker]{Pmaker()}}, \code{\link[=Smaker]{Smaker()}}
}
