% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.r
\name{ones}
\alias{ones}
\title{Ones Vector}
\usage{
ones(n)
}
\arguments{
\item{n}{how many ones}
}
\value{
a column vector of ones as an integer matrix
}
\description{
Make a column vector of ones.
}
\examples{
ones(5)
str(ones(5))
}

