% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMatrix.R
\name{plotMatrix}
\alias{plotMatrix}
\title{Plot a matrix}
\usage{
plotMatrix(A)
}
\arguments{
\item{A}{a matrix}
}
\value{
a ggplot object
}
\description{
plotMatrix is a R variant of Matlab's spy function.
}
\examples{

# the no-three-way interaction configuration
A <- hmat(c(3,3,3), subsets(1:3, 2))
plotMatrix(A)
plotMatrix(markov(A))

}
\author{
David Kahle \email{david.kahle@gmail.com}
}
