% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{pois_reg}
\alias{pois_reg}
\title{Fit a Poisson Regression model with algebraic methods}
\usage{
pois_reg(model, data, iter = 10000, burn = 1000, thin = 10,
  engine = c("Cpp", "R"), method = c("ipf", "mcmc"), moves,
  hit_and_run = FALSE, SIS = FALSE, non_uniform = FALSE, ...)
}
\arguments{
\item{model}{hierarchical poisson model specification}

\item{data}{data, as a data frame with raw data with discrete covariates}

\item{iter}{number of chain iterations}

\item{burn}{burn-in}

\item{thin}{thinning}

\item{engine}{C++ or R? (C++ yields roughly a 20-25x speedup)}

\item{method}{should the expected value (exp) be fit using
iterative proportional fitting (via loglin) or the MCMC as the
average of the steps?}

\item{moves}{the markov moves for the mcmc (as columns of a
matrix).}

\item{...}{...}

\item{init}{the initialization of the chain. by default, this is
the observed table}
}
\value{
a list containing named elements \itemize{ \item
  \code{steps}: an integer matrix whose columns represent
  individual samples from the mcmc. \item \code{moves}: the moves
  used for the proposal distribution in the mcmc, computed with
  4ti2 (note that only the positive moves are given). \item
  \code{acceptProb}: the average acceptance probability of the
  moves, including the thinned moves. \item \code{param}: the
  fitted parameters of the log linear model. \item \code{df}:
  parameters per term in the model \item \code{quality}: model
  selection statistics AIC, AICc, and BIC. \item 
  \code{residuals}: the (unstandardized) pearson residuals (O -
  E) / sqrt(E) \item \code{call}: the call. \item \code{obs}: the
  contingency table given. \item \code{exp}: the fit contingency
  table as an integer array. \item \code{A}: the sufficient
  statistics computing matrix (from Tmaker). \item 
  \code{p.value}: the exact p-values of individual tests,
  accurate to Monte-Carlo error.  these are computed as the
  proportion of samples with statistics equal to or larger than
  the oberved statistic. \item \code{mid.p.value}: the mid
  p.values, see Agresti pp.20--21. \item \code{statistic}: the
  pearson's chi-squared (X2), likelihood ratio (G2), 
  Freeman-Tukey (FT), Cressie-Read (CR), and Neyman modified
  chi-squared (NM) statistics computed for the table given. \item
  \code{sampsStats}: the statistics computed for each mcmc
  sample. \item \code{cells}: the number of cells in the table.
  \item \code{method}: the method used to estimate the table. }
  @importFrom stats model.frame
}
\description{
Fit a Poisson Regression model with algebraic methods
}
