% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/political_goals.R
\docType{data}
\name{political_goals}
\alias{political_goals}
\title{Relative Rankings of Importance of Four Political Goals}
\format{
An atomic named vector of length 4! = 24
}
\usage{
data(political_goals)
}
\description{
2262 Germans were asked to rank the following in order of importance : (1)
maintain order in the nation, "order"; (2) give people more say in decisions
in the government, "say"; (3) fight rising prices, "prices"; and (4) protect
freedom of speech, "freedom".
}
\references{
Barnes, S. H. and Kaase, M. (1979). \emph{Political Action: Mass
Participation in Five Western Countries}. Sage, Beverly Hills, CA.

Croon, M. A. (1998). Latent class models for the analysis of
rankings. \emph{New Developments in Psychological Choice Modeling},
99--121. Feger, Klauer, and de Soete, eds. North-Holland, Amsterdam.
}
\author{
The first reference
}
\keyword{datasets}
