\docType{data}
\name{politics}
\alias{politics}
\title{Politics by personality}
\format{A 2x2 (contingency) table}
\usage{
data(politics)
}
\description{
Politics by personality
}
\author{
David Kahle, simplifying the dataset found on p.622 of
Sheskin (see References)
}
\references{
Sheskin, D. J. Handbook of Parametric and Nonparametric
Statistical Procedures. 4ed. Chapman and Hall/CRC Press,
2007.
}
\keyword{datasets}

