\name{print.hierarchical}
\alias{print.hierarchical}
\title{Pretty printing of hierarchical's output.}
\usage{
\method{print}{hierarchical}(x, digits = 4, ...)
}
\arguments{
  \item{x}{an object of class hierarchical}

  \item{digits}{digits to round to}

  \item{...}{additional parameters}
}
\value{
Invisible string of the printed object.
}
\description{
Pretty printing of hierarchical's output.
}
\examples{
# see ?hierarchical
}

