% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.loglinear.r
\name{print.loglinear}
\alias{print.loglinear}
\title{Pretty printing of loglinear's output}
\usage{
\method{print}{loglinear}(x, digits = 4, ...)
}
\arguments{
\item{x}{an object of class loglinear}

\item{digits}{digits to round to}

\item{...}{additional parameters}
}
\value{
Invisible string of the printed object.
}
\description{
Pretty printing of loglinear's output.
}
\examples{

# see ?loglinear


}
\seealso{
loglinear
}

