% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.m2.r
\name{print.m2}
\alias{print.m2}
\title{Pretty printing of Macaulay2 output.}
\usage{
\method{print}{m2}(x, ...)
}
\arguments{
\item{x}{an object of class m2}

\item{...}{additional parameters}
}
\value{
Invisible string of the printed object.
}
\description{
Pretty printing of Macaulay2 output.
}
\examples{
\dontrun{

m2("13^1000")

}

}
