% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.polyOptim.r
\name{print.polyOptim}
\alias{print.polyOptim}
\title{Pretty printing of polyOptim (Bertini) output.}
\usage{
\method{print}{polyOptim}(x, lagrange = FALSE, digits = 3, ...)
}
\arguments{
\item{x}{an object of class polyOptim, bertini}

\item{lagrange}{show values of lagrange multipliers?}

\item{digits}{digits to round to}

\item{...}{additional parameters}
}
\value{
Invisible string of the printed object.
}
\description{
Pretty printing of polyOptim (Bertini) output.
}
\examples{

# see ?polyOptim


}
