\name{print.spectral}
\alias{print.spectral}
\title{Pretty printing of spectral's output.}
\usage{
\method{print}{spectral}(x, digits = 3, ...)
}
\arguments{
  \item{x}{an object of class spectral}

  \item{digits}{digits to round to}

  \item{...}{additional parameters}
}
\value{
Invisible string of the printed object.
}
\description{
Pretty printing of spectral's output.
}
\examples{
# see ?spectral
}

