% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.tableau}
\alias{print.tableau}
\title{Pretty printing of tableau output.}
\usage{
\method{print}{tableau}(x, ...)
}
\arguments{
\item{x}{an object of class tableau}

\item{...}{...}
}
\value{
Invisible string of the printed object.
}
\description{
Pretty printing of tableau output.
}
\examples{
# see ?tableau
}

