% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.latte}
\alias{read.latte}
\title{Read a latte file from disk}
\usage{
read.latte(file)
}
\arguments{
\item{file}{the name of the file}
}
\value{
an invisible form of the saved output.
}
\description{
\code{read.latte} reads a latte-formatted file a file.
}
\examples{
\dontrun{

mat <- matrix(sample(9), 3, 3)
mat
write.latte(mat, "foo.hrep")
file.show("foo.hrep")
read.latte("foo.hrep")

attr(mat, "linearity") <- c(1, 3)
attr(mat, "nonnegative") <- 2
mat
write.latte(mat, "foo.hrep")
file.show("foo.hrep")
read.latte("foo.hrep")

file.remove("foo.hrep")

}
}
\seealso{
\code{\link{read.latte}}
}

