% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab2vec.r, R/vec2tab.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{tab2vec}
\alias{reexports}
\alias{vec2tab}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{latter}{\code{\link[latter]{tab2vec}}, \code{\link[latter]{vec2tab}}}
}}

