\name{rvotes}
\alias{rvotes}
\title{Generate spectral data}
\usage{
rvotes(nVoters, nObjects, kSelected)
}
\arguments{
  \item{nVoters}{number of voters voting}

  \item{nObjects}{number of objects up for selection}

  \item{kSelected}{number of objects selected by each
  voter}
}
\value{
...
}
\description{
Generate spectral data
}
\examples{
rvotes(100, 10, 3)
}

