% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setBertiniPath}
\alias{setBertiniPath}
\title{Set Bertini Path}
\usage{
setBertiniPath(path)
}
\arguments{
\item{path}{a character string, the path to Bertini}
}
\value{
invisible bertiniPath
}
\description{
This function sets the Bertini path either by (1) passing it a character string or (2) using file.choose.
}
\examples{
\dontrun{

setBertiniPath()

}
}

