% Generated by roxygen2 (4.0.2): do not edit by hand
\name{setLattePath}
\alias{setLattePath}
\title{Set LattE Path}
\usage{
setLattePath(path)
}
\arguments{
\item{path}{a character string, the path to LattE (the function count, for example)}
}
\value{
invisible lattePath
}
\description{
This function sets the LattE path either by (1) passing it a character string or (2) using file.choose.
}
\examples{
\dontrun{

setLattePath()

}
}

