% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setM2Path}
\alias{setM2Path}
\title{Set Macaulay2 Path}
\usage{
setM2Path(path)
}
\arguments{
\item{path}{a character string, the path to m2}
}
\value{
invisible m2Path
}
\description{
This function sets the Macaulay2 path either by (1) passing it a character string or (2) using file.choose.
}
\examples{
\dontrun{

setM2Path()

}
}

