% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setMarkovPath}
\alias{setMarkovPath}
\title{Set 4ti2 Path}
\usage{
setMarkovPath(path)
}
\arguments{
\item{path}{a character string, the path to 4ti2 (the function markov, for example)}
}
\value{
invisible markovPath
}
\description{
This function sets the 4ti2 path either by (1) passing it a character string or (2) using file.choose.
}
\examples{
\dontrun{

setMarkovPath()

}
}

