% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.r
\name{setPaths}
\alias{setBertiniPath}
\alias{setLattEPath}
\alias{setLattePath}
\alias{setM2Path}
\alias{setMarkovPath}
\alias{setPaths}
\title{Set paths to external functions}
\usage{
setM2Path(path)

setBertiniPath(path)

setLattePath(path)

setMarkovPath(path)
}
\arguments{
\item{path}{a character string, the path to m2 (for Macaulay2), bertini (for Bertini), markov (say, for 4ti2), and count (say, for LattE)}
}
\value{
an invisible character string, the path found
}
\description{
These functions set the path to external programs either by (1) passing them a character string or (2) using \code{\link{file.choose}}.
}
\examples{

\dontrun{ # the below code requires suggested external software

## for Macaulay2
getOption("m2Path")
setM2Path() 

## for Bertini
getOption("bertiniPath")
setBertiniPath() 

## for LattE
getOption("lattePath")
setLattePath() 

## for 4ti2 (typically the same as LattE)
getOption("markovPath")
setMarkovPath() 



## each of these functions can be used statically as well
(markovPath <- getOption("markovPath"))
setMarkovPath("/path/to/4ti2/directory")
getOption("markovPath")
setMarkovPath(markovPath) # undoes example



}

}
\author{
David Kahle \email{david.kahle@gmail.com}
}

