\name{summary.bertini}
\alias{summary.bertini}
\title{Summarize Bertini output.}
\usage{
\method{summary}{bertini}(object, ...)
}
\arguments{
  \item{object}{an object of class bertini}

  \item{...}{additional parameters}
}
\value{
Invisible string of the printed object.
}
\description{
Summarize Bertini output.
}
\examples{
# see ?bertini
}

