% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{tab-fill}
\alias{tab-fill}
\alias{tab_fill}
\alias{tabFill}
\title{Fill a table with a number}
\usage{
tab_fill(tab, fill = 1L)

tabFill(...)
}
\arguments{
\item{tab}{a contingency table}

\item{fill}{the number to fill the contingency table with}

\item{...}{...}
}
\value{
a named numeric vector
}
\description{
Fill a table with a number
}
\examples{

Titanic
tab_fill(Titanic)
tab_fill(Titanic, 1L)
tab_fill(Titanic, 0L)

nCells <- prod(dim(Titanic))
tab_fill(Titanic, rpois(nCells, 5))

}
