% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tab2array.r
\name{tab2array}
\alias{tab2array}
\title{Table to array conversion}
\usage{
tab2array(tab)
}
\arguments{
\item{tab}{a table}
}
\value{
an array
}
\description{
Convert a table into an array.
}
\examples{
data(handy)
handy
tab2array(handy)

data(Titanic)
Titanic
tab2array(Titanic)
}
\seealso{
\code{\link{tab2vec}}, \code{\link{array2tab}}
}

