% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabFill.R
\name{tabFill}
\alias{tabFill}
\title{Fill a table with a number}
\usage{
tabFill(tab, fill = 1L)
}
\arguments{
\item{tab}{a contingency table}

\item{fill}{the number to fill the contingency table with}
}
\value{
a named numeric vector
}
\description{
Fill a table with a number
}
\examples{

Titanic
tabFill(Titanic)
tabFill(Titanic, 1)
tabFill(Titanic, 0)

nCells <- prod(dim(Titanic))
tabFill(Titanic, rpois(nCells, 5))


}
