% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{triangular}
\alias{triangular}
\alias{upper}
\alias{lower}
\title{Create a triangular matrix}
\usage{
upper(x)

lower(x)
}
\arguments{
\item{x}{a vector}
}
\value{
...
}
\description{
Create a upper or lower triangular matrix.
}
\examples{
upper(1:3)
lower(1:3)

upper(1:6)
lower(1:6)

upper(rnorm(6))

}
