% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makers.r
\name{Umaker}
\alias{Umaker}
\title{U matrix (rank data)}
\usage{
Umaker(m)
}
\arguments{
\item{m}{the number of objects}
}
\value{
...
}
\description{
Compute the generalized marginals matrix for a full ranking of m
objects.  Umaker generalized Mmaker.
}
\details{
This is the transpose of the generalized marginals matrix
presented in Marden (1995).
}
\examples{


data(politicalGoals)

lambdas <- apply(partitions(4), 1, function(v) v[v != 0])




}
\references{
Marden, J. I. (1995). \emph{Analyzing and Modeling
  Rank Data}, London: Chapman & Hall. pp.47--48.
}
\seealso{
\code{\link{Mmaker}}, \code{\link{Pmaker}},
  \code{\link{Smaker}}
}

