% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/latte.r
\name{write.latte}
\alias{write.latte}
\title{Write a latte file to disk}
\usage{
write.latte(mat, file)
}
\arguments{
\item{mat}{a matrix; for example the output of \code{\link{hmat}}}

\item{file}{the name of the file}
}
\value{
an invisible form of the saved output.
}
\description{
\code{write.latte} writes a latte-formatted file a file.
}
\examples{
\dontrun{

(mat <- matrix(sample(9), 3, 3))
write.latte(mat, "foo.hrep")
file.show("foo.hrep")
read.latte("foo.hrep")
read.latte("foo.hrep", "Ab")

attr(mat, "linearity") <- c(1, 3)
attr(mat, "nonnegative") <- 2
mat
write.latte(mat, "foo.hrep")
file.show("foo.hrep")
read.latte("foo.hrep")

file.remove("foo.hrep")

}
}
\seealso{
\code{\link{write.latte}}
}

