% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.r
\docType{package}
\name{algstat}
\alias{algstat}
\alias{algstat-package}
\alias{package-algstat}
\title{algstat : Algebraic statistics in R}
\description{
algstat is a package for algebraic statistics in R.  Current applications
include exact inference in log-linear models for contingency table data,
analysis of ranked and partially ranked data, and general purpose tools for
multivariate polynomials, building on the mpoly package.  To aid in the
process, algstat has ports to Macaulay2, Bertini, LattE and 4ti2.
}

