% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algstat-package.R
\docType{package}
\name{algstat}
\alias{algstat}
\alias{package-algstat}
\title{algstat: Algebraic Statistics in R}
\description{
algstat is a package for algebraic statistics in R.  Current applications
include exact inference in log-linear models for contingency table data,
analysis of ranked and partially ranked data, and general purpose tools for
multivariate polynomials, building on the mpoly package.  To aid in the
process, algstat leverages ports to Macaulay2 (through m2r), Bertini (through
bertini), LattE and 4ti2 (through latte).
}
