% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alm.R
\name{alm}
\alias{alm}
\alias{print.alm}
\title{Algebraic least squares regression}
\usage{
alm(formula, data, ...)

\method{print}{alm}(x, ...)
}
\arguments{
\item{formula}{A formula specifying the basis monomials, passed to
\link[stats:model.matrix]{stats::model.matrix}}

\item{data}{The dataset on which to apply the method, passed to
\link[stats:model.matrix]{stats::model.matrix}}

\item{...}{Additional parameters, presently discarded}
}
\value{
A numeric vector of model parameter estimates
}
\description{
Algebraic least squares regression
}
\examples{


## fitting a linear formula
########################################

n <- 101
df <- data.frame(x = seq(0, 1, length.out = n))
df$y <- 3 - 2*df$x + rnorm(n, 0, sd = .15)

plot(df)

(mod <- alm(~ x + y, data = df))
(y_coef <- coef(mod)[["y"]])
coef(mod) / y_coef

lm(y ~ x, data = df)


## fitting a quadratic formula
########################################

df <- rvnorm(100, mp("x^2 + (2 y)^2 - 1"), sd = .05, output = "tibble")
ggplot(df, aes(x, y)) + geom_point() + coord_equal()
p <- function(x, deg) {
  out <- poly(x, deg, raw = TRUE, simple = TRUE)
  names <- paste0(sprintf("\%s^", deparse(substitute(x))), 1:deg)
  names <- gsub("\\\\^1", "", names)
  colnames(out) <- names
  out
}
mod <- alm(~ p(x,2) + p(y,2) + I(x*y), data = df)
mod
str(mod)

poly <- paste(
  coef(mod), 
  c("1", "x", "x^2", "y", "y^2", "x y"), 
  collapse = " + "
)
poly <- mp(poly)
ggvariety(poly, xlim = c(-2, 2), ylim = c(-2, 2)) +
  geom_point(aes(x, y), data = df, alpha = .1, inherit.aes = FALSE) +
  coord_equal(xlim = c(-1.2, 1.2), ylim = c(-1.2, 1.2))
  


}
