% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alm.R
\name{alm}
\alias{alm}
\alias{print.alm}
\title{Algebraic least squares regression}
\usage{
alm(form, data, ...)

\method{print}{alm}(x, ...)
}
\arguments{
\item{form}{A formula specifying the basis monomials, passed to
\link[stats:model.matrix]{stats::model.matrix}}

\item{data}{The dataset on which to apply the method, passed to
\link[stats:model.matrix]{stats::model.matrix}}

\item{...}{Additional parameters, presently discarded}
}
\value{
A numeric vector of model parameter estimates
}
\description{
Algebraic least squares regression
}
\examples{

n <- 101
df <- data.frame(x = seq(0, 1, length.out = n))
df$y <- 3 - 2*df$x + rnorm(n, 0, sd = .15)

with(df, plot(x, y))

(mod <- alm(~ x + y, data = df))
-(mod / mod[["y"]])[-which(names(mod) == "y")]

lm(y ~ x, data = df)


}
