% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makers.r
\name{Amaker}
\alias{Amaker}
\title{Distance transitive matrix}
\usage{
Amaker(m, k)
}
\arguments{
\item{m}{the number of objects}

\item{k}{the number of objects selected}
}
\value{
...
}
\description{
Compute the distance transitive matrix for a survey in which k objects are selected from a group of m
}
\examples{
Amaker(4, 2)
}
\seealso{
\code{\link{Tmaker}}, \code{\link{Emaker}}, \code{\link{Mmaker}}, \code{\link{Pmaker}}, \code{\link{Smaker}}
}

