% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cookie.r
\docType{data}
\name{cookie}
\alias{cookie}
\title{Girl Scout Cookie Preferences}
\format{A list with elements "freq", "raw", "sorted", and "cookies"}
\usage{
data(cookie)
}
\description{
A small approval dataset on Girl Scout cookies.
}
\author{
Ann Johnston and Michael Orrison
}
\references{
Johnston, A. and Orrison, M. (2011) Markov Bases for Noncommutative Harmonic Analysis of Partially Ranked Data.  Undergraduate Thesis, Department of Mathematics, Harvey Mudd College.
}
\keyword{datasets}

