% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countFiber.R
\name{countFiber}
\alias{countFiber}
\alias{memCountFiber}
\title{Count the elements of a fiber Ax = b}
\usage{
countFiber(A, b, dir = tempdir(), opts = "", quiet = TRUE)

memCountFiber(A, b, dir = tempdir(), opts = "", quiet = TRUE)
}
\arguments{
\item{A}{the A matrix of Ax = b}

\item{b}{the b vector of Ax = b}

\item{dir}{directory to place the files in, without an ending /}

\item{opts}{options for count}

\item{quiet}{show latte output}

\item{...}{...}
}
\value{
an integer
}
\description{
Count the elements of a fiber Ax = b
}
\details{
\code{countFiber} uses LattE's count function (via algstat's 
\code{\link{count}} function) to count the fiber  In many cases,
the number of such tables is enormous.  In these cases, instead
of giving back an integer \code{countFiber} provides a character
string with the integer in it; see examples.
}
\examples{

\dontrun{ 


data(politics)
(A <- hmat(c(2,2), list(1, 2)))
countTables(politics, A)
b <- A \%*\% tab2vec(politics)
countFiber(A, b)




}

}
\seealso{
\code{\link{count}}
}

