% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{drugs}
\alias{drugs}
\title{Use of Alcohol, Cigarettes, and Marijuana by High School Students}
\format{A 2x2x2 (contingency) table}
\usage{
data(drugs)
}
\description{
A multi-way contingency table of alcohol, cigarettes, and marijuana use by high school students.
}
\author{
Wright State University School of Medicine and the United Health Services in Dayton, Ohio
}
\references{
Agresti, A. (2002). \emph{Categorical Data Analysis}, Basel: John Wiley & Sons, 2ed. (p.322)
}
\keyword{datasets}

