% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvariety.R
\name{ggvariety}
\alias{ggvariety}
\title{Plot a variety}
\usage{
ggvariety(mp, xlim = c(-1, 1), ylim = c(-1, 1), n = 101, nx = n, ny = n, ...)
}
\arguments{
\item{mp}{an mpoly object}

\item{xlim}{vector representing x bounds}

\item{ylim}{vector representing y bounds}

\item{n}{number of mesh points in each dimension}

\item{nx}{number of mesh points in the abcissa (x)}

\item{ny}{number of mesh points in the ordinate (y)}

\item{...}{additional parameters to pass to \code{\link[=geom_contour]{geom_contour()}}}
}
\value{
A ggplot object containing variety plot
}
\description{
Uses geom_contour() and ggplot() to plot an mpoly object representing a
variety in 2D space.
}
\examples{

## basic usage
##################################################

ggvariety("y - x")
ggvariety("y - x^2")
ggvariety("x^2 + y^2 - 1")
ggvariety(c("x^2 + y^2 - 1", "y - x"))



## setting limits
##################################################

ggvariety("y - x^2")
ggvariety("y - x^2", xlim = c(-2,2), ylim = c(-2,2))


## ggplot2 styling
##################################################

library("ggplot2")
ggvariety("x^2 + y^2 - 1") + coord_equal()
ggvariety("x^2 + y^2 - 1", color = "red") + coord_equal()
ggvariety("x^2 + y^2 - 1", size = 2) + coord_equal()
ggvariety("x^2 + y^2 - 1", size = 2, alpha = .2) + coord_equal()
ggvariety("x^2 + y^2 - 1", linetype = 2) + coord_equal()
ggvariety("x^2 + y^2 - 1") + coord_equal() + theme_bw()
ggvariety("x^2 + y^2 - 1") + coord_equal() + theme_classic()
ggvariety("x^2 + y^2 - 1") + coord_equal() + theme_void()

ggvariety(c("x^2 + y^2 - 1", "(x^2 + y^2)^3 - 4 x^2 y^2")) +
  coord_equal() + theme_void() +
  scale_color_manual(values = c("red", "blue"), guide = FALSE)



## possible issues
##################################################

# at a low level, ggvariety() uses grDevices::contourLines()
# to numerically detect zero crossings. this is an imperfect process,
# so you may see gaps where none exist. as a general strategy, upping
# the number of sampled points on the grid is recommended.
# the below are commented to cut check time; they run

# ggvariety("y^2 - x^3 - x^2") + coord_equal()
# ggvariety("y^2 - x^3 - x^2", n = 201) + coord_equal()
# ggvariety(mp(c("x^2 + y^2 - 1", "y - x")))

}
\author{
Phillip Hossu, Ryan Hebdon, Chong Sun, Grant Innerst, David Kahle
}
