% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/handy.r
\docType{data}
\name{handy}
\alias{handy}
\title{Handedness Data}
\format{A 2x2 (contingency) table}
\usage{
data(handy)
}
\description{
A small fictional dataset on handedness.
}
\author{
David Kahle
}
\keyword{datasets}

