% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lawrence.R
\name{lawrence}
\alias{lawrence}
\title{Compute the Lawrence lifting of a configuration matrix}
\usage{
lawrence(A)
}
\arguments{
\item{A}{the configuration matrix of the model}
}
\value{
an integer matrix
}
\description{
Compute the Lawrence lifting of a configuration matrix
}
\examples{
A <- hmat(c(2,2), 1:2)
lawrence(A)
}
\author{
David Kahle \email{david.kahle@gmail.com}
}
\seealso{
Aoki, Hara, and Takemura (2012), p. 61
}

