% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lawrence.R
\name{lawrence}
\alias{lawrence}
\title{Compute the Lawrence lifting of a configuration matrix}
\usage{
lawrence(A, extent = c("short", "tall"))
}
\arguments{
\item{A}{the configuration matrix of the model}

\item{extent}{short or tall version}
}
\value{
an integer matrix
}
\description{
Compute the Lawrence lifting of a configuration matrix
}
\examples{
A <- hmat(c(2,2), 1:2)
lawrence(A)
lawrence(A, extent = "tall")

\dontrun{ # requires 4ti2 installed
markov(A)
markov(lawrence(A))
markov(lawrence(A, extent = "tall"))
}
}
\author{
David Kahle \email{david.kahle@gmail.com}
}
\seealso{
Aoki, Hara, and Takemura (2012), pp. 61-62
}

