% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmat.R
\name{pmat}
\alias{pmat}
\title{Construct a Model Matrix for Poisson Regression}
\usage{
pmat(levels, facets)
}
\arguments{
\item{levels}{a vector containing the number of levels of each
variable}

\item{facets}{the facets generating the hierarchical model, a
list of vectors of variable indices}
}
\value{
a matrix
}
\description{
Determine the A matrix associated with a hierarchical model on a
contingency table for Poisson Regression.
}
