% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bertini.r
\name{print.bertini}
\alias{print.bertini}
\title{Pretty Printing of Bertini Output}
\usage{
\method{print}{bertini}(x, digits = 3, ...)
}
\arguments{
\item{x}{an object of class bertini}

\item{digits}{digits to round to}

\item{...}{additional parameters}
}
\value{
Invisible string of the printed object.
}
\description{
Pretty printing of Bertini output.
}
\examples{
\dontrun{

# see ?bertini

variety("x^2 + 1")
variety(c("x^2 + 1 + y","y"))

}

}

