% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print.hierarchical.r
\name{print.hierarchical}
\alias{print.hierarchical}
\title{Pretty Printing of Hierarchical's Output}
\usage{
\method{print}{hierarchical}(x, digits = 4, ...)
}
\arguments{
\item{x}{an object of class hierarchical}

\item{digits}{digits to round to}

\item{...}{additional parameters}
}
\value{
Invisible string of the printed object.
}
\description{
Pretty printing of hierarchical's output.
}
\examples{
# see ?hierarchical
}

