% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.spectral.R
\name{print.spectral}
\alias{print.spectral}
\title{Pretty Printing of Spectral's Output}
\usage{
\method{print}{spectral}(x, digits = 3, ...)
}
\arguments{
\item{x}{an object of class spectral}

\item{digits}{digits to round to}

\item{...}{additional parameters}
}
\value{
Invisible string of the printed object.
}
\description{
Pretty printing of spectral's output.
}
\examples{

# see ?spectral


}
\seealso{
\code{\link{spectral}}
}
