% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{project-onto}
\alias{project-onto}
\alias{project_onto}
\alias{project_onto_perp}
\title{Vector Projection}
\usage{
project_onto(A, x)

project_onto_perp(A, x)
}
\arguments{
\item{A}{a matrix}

\item{x}{a vector}
}
\value{
...
}
\description{
Project a vector onto the column space of a matrix or the orthogonal complement of the column
space of a matrix; the null space of A transpose.
}
\examples{

A <- diag(5)[,1:2]
x <- 1:5
project_onto(A, x)
project_onto_perp(A, x)

}
\seealso{
\link[base:qr.fitted]{base::qr.fitted}
}
