% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r, R/plotMatrix.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{cpp_object_initializer}
\alias{tab2vec}
\alias{vec2tab}
\alias{\%>\%}
\alias{plot_matrix}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{Rcpp}{\code{\link[Rcpp]{cpp_object_initializer}}}

  \item{latte}{\code{\link[latte]{plot_matrix}}, \code{\link[latte]{plot_matrix}}, \code{\link[latte]{tab2vec}}, \code{\link[latte]{vec2tab}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}
}}

