% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvnorm.R
\name{rvnorm}
\alias{rvnorm}
\title{The Variety Normal Distribution}
\usage{
rvnorm(n, poly, sd, output = "simple", normalized = TRUE,
  chains = 4L, cores = getOption("mc.cores", 1L),
  warmup = floor(n/2), keep_warmup = FALSE, inject_direct = FALSE,
  verbose = FALSE, w, vars, numerator, denominator, refresh, ...)
}
\arguments{
\item{n}{The number of draws desired from each chain.}

\item{poly}{An mpoly object.}

\item{sd}{The "standard deviation" component of the normal kernel.}

\item{output}{\code{"simple"}, \code{"more"}, \code{"stanfit"}.}

\item{normalized}{If \code{TRUE}, the polynomial is gradient-normalized. This
is highly recommended.}

\item{chains}{The number of chains to run for the random number generation,
see \code{\link[=stan]{stan()}}.}

\item{cores}{The number of CPU cores to distribute the chains across, see
\code{\link[=stan]{stan()}}.}

\item{keep_warmup}{If \code{TRUE}, the MCMC warmup steps are included in the
output.}

\item{inject_direct}{Directly specify printed polynomial to string inject
into the stan code. Requires you specify \code{vars}, \code{numerator}, and
\code{denominator}.}

\item{verbose}{\code{TRUE} or \code{FALSE}; determines level of messaging.}

\item{w}{A box window (-w,w) of the same dimension as the number of
variables.}

\item{vars}{A character vector of the indeterminates in the distribution.}

\item{numerator, denominator}{A character(1) containing the printed numerator
of the variety normal distribution.}

\item{refresh}{The \code{refresh} argument of \code{\link[=stan]{stan()}}, which governs how
much information is provided to the user while sampling.}

\item{...}{Additional parameters to pass to \code{\link[=stan]{stan()}}.}
}
\value{
Either (1) matrix whose rows are the individual draws from the
distribution, (2) a \linkS4class{tbl_df} object with the draws along with
additional information, or (3) an object of class \linkS4class{stanfit}.
}
\description{
Un-normalized density function and random generation for the variety normal
distribution with mean equal to \code{poly} and "standard deviation" equal to
\code{sd}. Please see details for caveats.
}
\details{
If the variety you are interested in is connected, this strategy should work
well out of the box.  If it isn't, you'll likely need to rely on running
multiple chains, and it is very likely, if not probable, that the sampling
will be biased to one or more of those components and down-sample others.
Question: what is the relative likelihood of each component, or an equal unit
of length, on different components? How does this generalize to more
varieties of varying dimensions?
}
\examples{

\dontrun{ contains runs rstan

## basic usage
########################################

p <- mp("x^2 + y^2 - 1")
samps <- rvnorm(2000, p, sd = .1)
head(samps)

samps <- rvnorm(2000, p, sd = .1, output = "tibble")
ggplot(samps, aes(x, y)) + geom_point() + coord_equal()


}

}
\author{
David Kahle
}
