% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rvotes}
\alias{rvotes}
\title{Random Spectral Data}
\usage{
rvotes(nVoters, nObjects, kSelected)
}
\arguments{
\item{nVoters}{number of voters voting}

\item{nObjects}{number of objects up for selection}

\item{kSelected}{number of objects selected by each voter}
}
\value{
...
}
\description{
Generate spectral data for testing purposes.
}
\examples{
rvotes(100, 10, 3)
}

