% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.r
\name{setPaths}
\alias{set4ti2Path}
\alias{setBertiniPath}
\alias{setLattEPath}
\alias{setLattePath}
\alias{setPaths}
\title{Set paths to external functions}
\usage{
setBertiniPath(path)

setLattePath(path)

set4ti2Path(path)
}
\arguments{
\item{path}{a character string, the path to bertini (for
Bertini), markov (say, for 4ti2), and count (say, for LattE)}
}
\value{
an invisible character string, the path found
}
\description{
These functions set the path to external programs either by (1) 
passing them a character string or (2) using 
\code{\link{file.choose}}.
}
\examples{

\dontrun{ # the below code requires suggested external software


## for Bertini
getOption("bertini")
setBertiniPath() 

## for LattE
getOption("latte")
setLattePath() 

## for 4ti2 (typically the same as LattE)
getOption("4ti2")
set4ti2Path() 



## each of these functions can be used statically as well
(`4ti2Path` <- getOption("4ti2"))
set4ti2Path("/path/to/4ti2/directory")
getOption("4ti2")
set4ti2Path(`4ti2Path`) # undoes example



}

}
\author{
David Kahle \email{david.kahle@gmail.com}
}

