% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summary.bertini}
\alias{summary.bertini}
\title{Summarize Bertini Output}
\usage{
\method{summary}{bertini}(object, ...)
}
\arguments{
\item{object}{an object of class bertini}

\item{...}{additional parameters}
}
\value{
Invisible string of the printed object.
}
\description{
This function summarizes the output from Bertini.
}
\examples{
# see ?bertini
}

